+++ packages/youtube_player_flutter/example/lib/main.dart
@@ -3,7 +3,7 @@ import 'dart:developer';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:youtube_player_flutter/youtube_player_flutter.dart';
+import 'package:youtube_player_flutter_quill/youtube_player_flutter_quill.dart';
 
 import 'video_list.dart';