+++ packages/youtube_player_flutter/lib/src/widgets/playback_speed_button.dart
@@ -53,7 +53,7 @@ class _PlaybackSpeedButtonState extends State<PlaybackSpeedButton> {
         child: widget.icon ??
             Image.asset(
               'assets/speedometer.webp',
-              package: 'youtube_player_flutter',
+              package: 'youtube_player_flutter_quill',
               width: 20.0,
               height: 20.0,
               color: Colors.white,